requirements.txt
google-genai
langchain
streamlit